优化createRouterFunc 方法中的反射调用，提升单位时间内的执行次数 

在性能测试操作中，需要分别注释掉新老方法中的   r.error = r.Parse(xxxx），新老方法都调用同一个request的Parse方法，这部分可以注释掉，再执行测试

测试指令   go test -bench='Benchmark(Original|Optimized)' -run=none -benchmem -benchtime=2s -count=3
性能测试效果 
 
goos: windows
goarch: amd64
pkg: github.com/gogf/gf/v2/net/ghttp
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkOriginal-8      3957397               608.7 ns/op           176 B/op          5 allocs/op
BenchmarkOriginal-8      3910057               605.9 ns/op           176 B/op          5 allocs/op
BenchmarkOriginal-8      3944840               604.4 ns/op           176 B/op          5 allocs/op
BenchmarkOptimized-8     4023745               595.7 ns/op           176 B/op          5 allocs/op
BenchmarkOptimized-8     4016392               597.0 ns/op           176 B/op          5 allocs/op
BenchmarkOptimized-8     4019696               596.3 ns/op           176 B/op          5 allocs/op
PASS
ok      github.com/gogf/gf/v2/net/ghttp 18.075s

BenchmarkOptimized为优化后 的方法，BenchmarkOriginal为原方法

其中一次测试截图
![image](https://github.com/user-attachments/assets/aaeb8353-8dc2-4fdf-8ab6-b1031b11eaa4)

测试结果中，可以看到   第二项耗时提升7ns/op 左右，第三四项一样，  第一项指标表示在2秒时间内的执行次数 每轮对比都多出5万次以上

在每次操作的平均时间、每次操作分配的内存和每次操作的内存分配次数都差不多的情况下，执行次数越多，通常可以认为性能越好。执行次数的增加直接反映了代码在单位时间内能够处理更多的操作，从而提升系统的吞吐量和效率