@wln32 2024-09-11
@Royal-go 看起来提升不大，你可以考虑把HelloReq这样的参数缓存起来，不必每次都重新创建对象，减少gc，重复利用，记得每次用完需要清空