@wln32 2024-09-13
> > @Royal-go 看起来提升不大，你可以考虑把HelloReq这样的参数缓存起来，不必每次都重新创建对象，减少gc，重复利用，记得每次用完需要清空
> 
> 这里主要提升的是单位时间内的执行效率，也就是第一项指标，这个能提升系统吞吐量，HelloReq 只是模拟测试环境用，真正测的 是createRouterFunc 里面的反射处理

提升太小了，你要是600ns到500ns还可以考虑